% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_extraction.R
\name{adapter_dis}
\alias{adapter_dis}
\title{adapter_dis}
\usage{
adapter_dis(data, UMI_len = 10, flank = 1)
}
\arguments{
\item{data}{The output from the barcode match step. Should be a dataframe contains edit distance and
the adapter sequence}

\item{flank}{The length of flank when extract UMI, which is used to be tolerant of indels and dels.}
}
\value{
A dataframe with two columns, the first is the edit distance and the secons is
how many adapters have such an edit distance to the original adapter
}
\description{
Calculate the needleman score between original adapter and adapter in reads.
}
\details{
Calculate the edit distance between original adapter and adapter in reads to represent the data quality.
Only adapters upstream confidently identified barcodes will be used.
}
