% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(
  gtf_path = NULL,
  gene_bed_path = NULL,
  work_dir = "./",
  overwrite = TRUE,
  bed_chr_col = "chr",
  bed_start_col = "start",
  bed_end_col = "end",
  bed_strand_col = "strand",
  bed_gene_col = "gene"
)
}
\arguments{
\item{gtf_path}{The path for the gtf annotation.}

\item{gene_bed_path}{The path for the gene bed annotation.}

\item{overwrite}{The flag to indicate if the annotation file needs to be rewritten.}

\item{bed_chr_col}{The name of the column indicating the chromosome in the gene bed}

\item{bed_start_col}{The name of the column indicating the start position of the gene in the gene bed}

\item{bed_end_col}{The name of the column indicating the end position of the gene in the gene bed}

\item{bed_strand_col}{The name of the column indicating the gene strand in the gene bed}

\item{bed_gene_col}{The name of the column indicating the gene name in the gene bed}
}
\description{
Build annotations given if there already exists an annotation
}
\details{
Build annotations given if there already exists an annotation
}
