% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{bamGeneCoverage}
\alias{bamGeneCoverage}
\title{bamGeneCoverage}
\usage{
bamGeneCoverage(bam, gene_range_bed, outdir, bedtools = "bedtools")
}
\arguments{
\item{bam}{The filename of the input bam.}

\item{gene_range_bed}{The filename of the gene bed file, each row records the start and end position for the gene}

\item{outdir}{The folder to store the output file.}

\item{bedtools}{The path for the bedtools.}
}
\value{
A bed file including all genes without read alignment in the bam file.
}
\description{
detected genes with alignment in the bam file.
}
\details{
detected genes with alignment in the bam file.
}
