% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeMatch.R
\name{BarcodeFilter}
\alias{BarcodeFilter}
\title{BarcodeFilter}
\usage{
BarcodeFilter(data, mean_edit_thresh = 1.5)
}
\arguments{
\item{data}{The output from the barcode match step. Should be a dataframe contains the read name, cell barcode and its matching edit distance}

\item{mean_edit_thresh}{The maximum tolerance of mean edit distance for a cell barcode to be preserved}
}
\description{
Filter out barcode match with low quality
}
\details{
Filter out barcode match with low quality according to the mean edit distance of the cell barcode
}
