% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeMatch.R
\name{BarcodeMatch}
\alias{BarcodeMatch}
\title{BarcodeMatch}
\usage{
BarcodeMatch(
  seq,
  barcodes,
  mu = 20,
  sigma = 10,
  sigma_start = 10,
  k = 6,
  batch = 100,
  top = 5,
  cos_thresh = 0.25,
  alpha = 0.05,
  edit_thresh = 3,
  mean_edit_thresh = 1.5,
  UMI_len = 10,
  UMI_flank = 1,
  cores = 1
)
}
\arguments{
\item{seq}{A vector of strings recording the tag regions for each sequence.}

\item{barcodes}{A character vector of cell barcode}

\item{mu}{The mean end location of cell barcode in the softclip, using the middle sites of the softclip as default}

\item{sigma}{The start variance for the end position distribution}

\item{sigma_start}{The variance to buffer at first to avoid sudden convergence of variance}

\item{k}{The length of sub string match between softclips and barcodes,
only softclips and barcodes with high enough kmer match will be further validated}

\item{batch}{The number of softclips to search in one run}

\item{top}{Barcodes with top n cos similarity will be used for further validation}

\item{cos_thresh}{The minimum threshold of the cos similariy to preserve the cell barcode for further validation}

\item{alpha}{The probability thresh to filter out barcode match, matches
with too high edit distance or too far position have low probability to be correct}

\item{edit_thresh}{The highest tolerance of edit distance for the barcode match}

\item{mean_edit_thresh}{The maximum tolerance of mean edit distance for a cell barcode to be preserved}

\item{UMI_len}{The length of the UMI sequence}

\item{UMI_flank}{The length of the flank to add to tolerate the insertions and deletions around UMI}

\item{cores}{The number of cores to use for parallelization}
}
\description{
BarcodeMatch with parallelization
}
\details{
BarcodeMatch for large dataset which is paralleled by future.apply.
Large data will be split according to cores. The barcode k-mer index is built
once before the parallel loop and passed to each worker to avoid redundant work.
}
