% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_corres.R
\name{bins2iso}
\alias{bins2iso}
\title{bins2iso}
\usage{
bins2iso(start, end, sep = ",", split = "|")
}
\arguments{
\item{start}{A numerical vector storing the start positions for all exons in the isoform}

\item{end}{A numerical vector storing the end positions for all exons in the isoform}

\item{sep}{The character to split the start and end position for each exon in the isoform.}

\item{split}{The character to split the exons in the isoform}
}
\value{
A string representing isoform
}
\description{
Paste the exon bins (start and end) together as an isoform sequence
}
