% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{binsum}
\alias{binsum}
\title{binsum}
\usage{
binsum(bin, start_col = "start", end_col = "end")
}
\arguments{
\item{bin}{The bin matrix}

\item{start_col, end_col}{The name of the column recording the start/end position}
}
\value{
A number as the total length.
}
\description{
calculate the total length of a series of bins
}
\details{
calculate the total length of a series of bins
}
