% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{cell_iso_count_impute}
\alias{cell_iso_count_impute}
\title{cell_iso_count_impute}
\usage{
cell_iso_count_impute(data)
}
\arguments{
\item{data}{The input dataframe, including the reads, reads count and its aligned annotated isoform.}
}
\description{
imputate the isoform count for one gene in a cell.
}
\details{
Some truncated reads can have ambigous alignment to multiple annotated isoforms, this function will
iterative imputate the isoform count based on non-ambigous alignment.
}
