% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reads_filter.R
\name{cells_isoforms_size_filter}
\alias{cells_isoforms_size_filter}
\title{cells_isoforms_size_filter}
\usage{
cells_isoforms_size_filter(
  cell_isoform_table,
  ratio = 0.1,
  ...,
  thresh = 10,
  eps = 10
)

cells_isoforms_size_filter(
  cell_isoform_table,
  ratio = 0.1,
  ...,
  thresh = 10,
  eps = 10
)
}
\arguments{
\item{cell_isoform_table}{A dataframe for reads including its cell source,
isoform sequence, splicing sites and umi cluster size.}

\item{ratio}{The estimated frequency for UMIs which are enriched of sequencing errors and
couldn't be clustered correctly and become singletons}
}
\value{
A dataframe with isoform count after cluster size filtering for each cell

A dataframe with isoform count after cluster size filtering for each cell
}
\description{
Apply cluster size filtering to each isoform in each cell.

Apply cluster size filtering to each isoform in each cell.
}
