% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{cells_mid_filter}
\alias{cells_mid_filter}
\title{cells_mid_filter}
\usage{
cells_mid_filter(cells, cluster, isoform_mid)
}
\arguments{
\item{cells}{The string vector indicating the cell}

\item{cluster}{The factor vector indicating the UMI cluster}

\item{isoform_mid}{The string vector indicating middle splicing sites.}
}
\value{
A string vector indicating preserved patterns of middle splicing sites
}
\description{
Filter out middle splice sites pattern caused by wrong mapping.
}
\details{
Filter out middle splice sites pattern caused by wrong mapping given the information from UMI cluster as
reads within the same UMI cluster should come from the same transcript.
}
