% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{clean_file}
\alias{clean_file}
\title{clean_file}
\usage{
clean_file(dir)
}
\arguments{
\item{dir}{The output path for the LongcellPre pipeline}
}
\description{
clean the intermediate files for LongcellPre pipeline.
}
