% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{coreDetect}
\alias{coreDetect}
\title{coreDetect}
\usage{
coreDetect(cores)
}
\arguments{
\item{cores}{Number of cores used to do parallization.}
}
\description{
Check if the running platform has required number cores
}
\details{
Check if the running platform has required number cores, if not will use the maximum
}
