% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{createAnnotation}
\alias{createAnnotation}
\title{createAnnotation}
\usage{
createAnnotation(
  gtf_path = NULL,
  gene_bed_path = NULL,
  work_dir = "./",
  bed_chr_col = "chr",
  bed_start_col = "start",
  bed_end_col = "end",
  bed_strand_col = "strand",
  bed_gene_col = "gene"
)
}
\arguments{
\item{gtf_path}{The path for the gtf annotation.}

\item{gene_bed_path}{The path for the gene bed annotation.}

\item{bed_chr_col}{The name of the column indicating the chromosome in the gene bed}

\item{bed_start_col}{The name of the column indicating the start position of the gene in the gene bed}

\item{bed_end_col}{The name of the column indicating the end position of the gene in the gene bed}

\item{bed_strand_col}{The name of the column indicating the gene strand in the gene bed}

\item{bed_gene_col}{The name of the column indicating the gene name in the gene bed}

\item{word_dir}{The path to save the output}
}
\value{
A list with two dataframes, the first one is the gene bed and the second one is the gtf exon annotation
}
\description{
Build annotation based on the gtf or gene bed input
}
\details{
Build annotation based on the gtf or gene bed input
}
