% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{createSNN}
\alias{createSNN}
\title{createSNN}
\usage{
createSNN(dis, count = NULL, self = FALSE)
}
\arguments{
\item{dis}{The edge table from a UMI graph with two columns, each col represents a node.}

\item{count}{The weight for each node, if NULL, all weight will be set as 1.}

\item{self}{For the node with self loop in the graph, decide if themselves should be considered in the neibor counting.}

\item{iso_thresh}{The maximum threshold that two isoforms can be different in overlapping bases.}
}
\description{
count the number of shared neighbors between two connected nodes in a UMI graph
}
\details{
The function receives the edge table as input, count the shared neighbors between two connected nodes
}
