% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{dataSplit}
\alias{dataSplit}
\title{dataSplit}
\usage{
dataSplit(data, cores)
}
\arguments{
\item{data}{An input dataframe or vector}

\item{cores}{The number of cores required for parallel}
}
\value{
A list including split dataframe or vector
}
\description{
Split a dataframe by row or a vector uniformly given the number of cores
}
\details{
Split a dataframe by row or a vector uniformly given the number of cores
}
