% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{disagree_sites}
\alias{disagree_sites}
\title{disagree_sites}
\usage{
disagree_sites(from, to)
}
\arguments{
\item{from, to}{A string vector of middle splicing sites. From is the original middle splicing
sites, to is the middle splicing sites the from mapped to within the same UMI cluster}
}
\value{
A dataframe, recording the number of splicing sites existing in the correctly mapped
and wrongly mapped reads.
}
\description{
Identify the different sites between different isoforms within the same
UMI cluster
}
