% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{exon_id}
\alias{exon_id}
\title{exon_id}
\usage{
exon_id(strand, count)
}
\arguments{
\item{strand}{The gene strand, using "+" to indicate forward and using "-" to indicate reverse}
}
\value{
A character vector representing the exon id
}
\description{
Assign an id for each exon part in the gene bed
}
\details{
Assign an id for each exon part in the gene bed according to the gene strand, forward strand gene will
have exon id from smaller coordinates to higher, vice versa for reverse strand gene
}
