% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_output.R
\name{extractAnnotFromQname}
\alias{extractAnnotFromQname}
\title{extractAnnotFromQname}
\usage{
extractAnnotFromQname(qname, annot = "cell")
}
\arguments{
\item{qname}{A vector of string which stores the read name.}

\item{annot}{The prefix of the annotation to be extracted}
}
\value{
A vector of string which stores the annotations.
}
\description{
Extract the annotation for each read from the read name.
}
