% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_output.R
\name{extractReadSeq}
\alias{extractReadSeq}
\title{extractReadSeq}
\usage{
extractReadSeq(seq, start, end, exon_bins)
}
\arguments{
\item{seq}{The string of the parent sequence}

\item{start, end}{The start and the end position of the parent sequence in the genome}

\item{exon_bins}{A matrix with two columns, each row is a bin and each element represents the start and end position of the bin.}
}
\value{
A string of the extracted sequence.
}
\description{
extract the sequence from the parent sequence given the coordinates.
}
