% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_output.R
\name{extractSeq}
\alias{extractSeq}
\title{extractSeq}
\usage{
extractSeq(genome, chr, isoforms, strand, split = "|", sep = ",")
}
\arguments{
\item{genome}{The name of the genome, will be used to retrieve the genome object from BSgenome.}

\item{chr}{The chromosome index of the sequence to be extracted.}

\item{isoforms}{A string vector to indicate the isoforms.}

\item{strand}{The strand of the sequence to be extracted.}

\item{split}{The character to seperate the exons in the isoform string.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform string.}
}
\value{
A string of the extracted sequence.
}
\description{
extract the sequence from the genome given the splicing sites.
}
