% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{fastqMap}
\alias{fastqMap}
\title{fastqMap}
\usage{
fastqMap(
  fastq,
  out_name,
  genome_path,
  bed_path = NULL,
  minimap2 = "minimap2",
  samtools = "samtools",
  minimap2_thread = 4,
  samtools_thread = 4
)
}
\arguments{
\item{fastq}{The filename of the input fastq}

\item{out_name}{The filename of the output polished fastq}

\item{genome_path}{The filename of the reference genome}

\item{bed_path}{The filename of the reference bed for annotated splicing sites if provided,
BED12 file can be converted from GTF/GFF3 with ‘paftools.js gff2bed anno.gtf’.}

\item{minimap2}{The path of the minimap2}

\item{samtools}{The path of the samtools}

\item{minimap2_thread}{The number of threads for minimap2 parallization}

\item{samtools_thread}{The number of threads for samtools parallization}
}
\description{
Map the fastq file to genome via minimap2
}
\details{
Map the fastq file to genome via minimap2
}
