% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_filter.R
\name{frNN_dis}
\alias{frNN_dis}
\title{frNN_dis}
\usage{
frNN_dis(dis, n, eps = 10, if_direct = FALSE)
}
\arguments{
\item{dis}{A distance long table with 3 columns, each means node1, node2 and distance between them}

\item{n}{The total count of the nodes}

\item{eps}{The maximum threshold of distance to be preserved}

\item{if_direct}{If there is a direction between each nodes}
}
\value{
A frNN object
}
\description{
Build frNN object from distance long table.
}
\details{
Build frNN object from distance long table for cluster by dbscan package.
}
