% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_count_functions.R
\name{gene_umi_count}
\alias{gene_umi_count}
\title{gene_umi_count}
\usage{
gene_umi_count(
  cell_exon,
  qual,
  strand,
  bar = "barcode",
  isoform = "isoform",
  polyA = "polyA",
  sim_thresh = 7,
  split = "|",
  sep = ",",
  splice_site_thresh = 10,
  verbose = FALSE
)
}
\arguments{
\item{cell_exon}{the input dataframe which includes cell barcode, UMI sequence, isoform and polyA status.}

\item{qual}{The adapter distance from barcode matching to represent the data quality}

\item{bar}{The name of the column to represent cell barcode.}

\item{isoform}{The name of the column to represent the isoform.}

\item{polyA}{The name of the column to represent the polyA status.}

\item{sim_thresh}{the minimum threshold of needleman score to connect to different UMIs.}

\item{split}{The character to seperate different exons in the isoform representation.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform representation.}

\item{splice_site_thresh}{The minimum threshold of occurance  for an splice site to be preserved}

\item{verbose}{if print information during UMI clustering for each cell.}

\item{gene_cells_cluster}{The UMI cluster in each cell for a gene}

\item{relation}{The function to describe the relationship between the mean and variance of UMI cluster size}

\item{alpha}{the alpha of confidence interval for the variance of UMI cluster size.}
}
\description{
get the umi count for each cell in a gene.
}
\details{
get the umi count for each cell in a gene, the process includes clustering for UMI and correction for the isoform.
}
