% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{genes_distribute}
\alias{genes_distribute}
\title{genes_distribute}
\usage{
genes_distribute(data, cores, gene_col = "gene")
}
\arguments{
\item{data}{The input dataframe, each row is a read from a gene.}

\item{cores}{The number of cores required for parallel}

\item{gene_col}{The name of the column in the input data which record the gene name}
}
\value{
A list including split dataframe
}
\description{
Distribute genes to different cores to make uniform mission distribution
}
\details{
Distribute genes to different cores given their number of reads to let each
core receive similar number of reads
}
