% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{graph_to_cluster}
\alias{graph_to_cluster}
\title{graph_to_cluster}
\usage{
graph_to_cluster(graph_list)
}
\arguments{
\item{graph_list}{a list of igraph object, each object represents a cluster.}
}
\description{
extract the cluster information for nodes from a list of sub-graphs.
}
\details{
each sub-graph in the list represents a cluster, this function assigns the cluster id
for nodes within those sub-graphs.
}
