% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{gtf2bed}
\alias{gtf2bed}
\title{gtf2bed}
\usage{
gtf2bed(gtf_path, out_path, if_store_binary = TRUE)
}
\arguments{
\item{gtf_path}{The path of input gtf files}

\item{out_path}{The path to store the output}

\item{if_store_binary}{Whether to save the rds format of gene bed}
}
\description{
Transform gtf annotation to gene bed annotation
}
\details{
Exons annotated in gtf will be split into non-overlapping parts and saved in gene bed annotation
}
