% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{init}
\alias{init}
\title{init}
\usage{
init(work_dir)
}
\arguments{
\item{work_dir}{The specified directory for the output of LongcellPre}
}
\description{
Initiation of the LongcellPre pipeline
}
\details{
Build the directories for the output of LongcellPre
}
