% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{intron_only}
\alias{intron_only}
\title{intron_only}
\usage{
intron_only(
  reads,
  gtf,
  gtf_start_col = "start",
  gtf_end_col = "end",
  sep = ",",
  split = "|"
)
}
\arguments{
\item{reads}{A string vector contains strings for reads}

\item{gtf}{The gtf annotation, each row is an exon for an isoform}

\item{gtf_start_col}{the name of the column which stores the start position of exon in the gtf.}

\item{gtf_end_col}{the name of the column which stores the end position of exon in the gtf.}

\item{sep}{The character to split the start and end position for each exon in the isoform.}

\item{split}{The character to split the exons in the isoform}
}
\value{
A boolean vector.
}
\description{
judge if a read only contains intron sequence
}
\details{
This function compares the coverage between the read and the annotated isoform, if the read
has no overlap with the annotated isoform, this read is classified as intron.
}
