% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{iso_corres}
\alias{iso_corres}
\title{iso_corres}
\usage{
iso_corres(
  transcripts,
  gene,
  gtf,
  thresh = 3,
  overlap_thresh = 0.25,
  end_bias = 200,
  gtf_gene_col = "gene",
  gtf_iso_col = "transname",
  gtf_start_col = "start",
  gtf_end_col = "end",
  sep = ",",
  split = "|"
)
}
\arguments{
\item{transcripts}{A string vector storing reads}

\item{gene}{The gene name, should exist in the gtf annotation}

\item{gtf}{The gtf annotation, each row is an exon for an isoform}

\item{thresh}{The maximum threshold for the total offset of middle splicing sites.}

\item{overlap_thresh}{The minimum threshold for the coverage of the annotated isoform.}

\item{end_bias}{The maximum threshold for the offset of start and end position.}

\item{gtf_gene_col}{the name of the column which stores gene name in the gtf.}

\item{gtf_iso_col}{the name of the column which stores isoform name in the gtf.}

\item{gtf_start_col}{the name of the column which stores the start position of exon in the gtf.}

\item{gtf_end_col}{the name of the column which stores the end position of exon in the gtf.}

\item{sep}{The character to split the start and end position for each exon in the isoform.}

\item{split}{The character to split the exons in the isoform}
}
\value{
A dataframe with two columns, the first column records the reads, the second column records the isoform
the read could align to.
}
\description{
map the reads to the annotated isoforms for one gene
}
\details{
map the reads to the annotated isoforms for one gene
}
