% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{iso_count_impute}
\alias{iso_count_impute}
\title{iso_count_impute}
\usage{
iso_count_impute(
  data,
  cell_col = "cell",
  iso_col = "isoform",
  overlap_col = "overlap",
  count_col = "count"
)
}
\arguments{
\item{data}{The input dataframe, including the reads, reads count and its aligned annotated isoform.}

\item{cell_col}{The name of the column in the input data which records the cell barcode.}

\item{iso_col}{The name of the column in the input data which records the isoform.}

\item{overlap_col}{The name of the column in the input data which records the ratio of coverage.}

\item{count_col}{The name of the column in the input data which records the reads count.}
}
\description{
imputate the isoform count for one gene in multiple cells.
}
\details{
Some truncated reads can have ambigous alignment to multiple annotated isoforms, this function will
iterative imputate the isoform count based on non-ambigous alignment.
}
