% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{iso_end_diff}
\alias{iso_end_diff}
\title{iso_end_diff}
\usage{
iso_end_diff(read1, read2, split = "|", sep = ",")
}
\arguments{
\item{read1, read2}{A string or matrix to record the exons in a read}

\item{split}{The character to split the exons in the isoform}

\item{sep}{The character to split the start and end position for each exon in the isoform.}
}
\value{
Two numerical number to denote the difference and the 5 and 3 end.
}
\description{
calculate the difference at the two ends in the alignment between two reads in base pair
}
