% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{iso_full_dis}
\alias{iso_full_dis}
\title{iso_full_dis}
\usage{
iso_full_dis(read1, read2, split = "|", sep = ",")
}
\arguments{
\item{read1, read2}{A string or matrix to record the exons in a read}

\item{split}{The character to split the exons in the isoform}

\item{sep}{The character to split the start and end position for each exon in the isoform.}
}
\value{
A numerical number to denote the difference.
}
\description{
calculate the difference in the alignment between two reads in base pair
}
