% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{iso_len}
\alias{iso_len}
\title{iso_len}
\usage{
iso_len(iso, sep = ",", split = "|")
}
\arguments{
\item{iso}{The input string of the isoform}

\item{sep}{The character to seperate the splicing sites in an exon}

\item{split}{The character to seperate the exon in the isoform}
}
\value{
The length of the isoform
}
\description{
calculate the length of the isoform
}
