% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{isoform_correct}
\alias{isoform_correct}
\title{isoform_correct}
\usage{
isoform_correct(gene_isoform, preserve_mid)
}
\arguments{
\item{gene_isoform}{A data frame containing cell, gene,and isoform informaion including
start, middle splicing sites, end position and polyA existence}

\item{preserve_mid}{A string vector indicating preserved patterns of middle splice sites}
}
\value{
A dataframe including the corrected start, middle splicing sites, end position, polyA existence
and the number of reads of each UMI cluster for each gene in each cell
}
\description{
Corret for wrong mapping and truncation for each UMI cluster in each gene in each cell
}
\details{
Corret for wrong mapping and truncation for each UMI cluster in each gene in each cell
}
