% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_count_functions.R
\name{isoform_correct_filter}
\alias{isoform_correct_filter}
\title{isoform_correct_filter}
\usage{
isoform_correct_filter(
  gene_cells_cluster,
  filter_ratio,
  strand,
  split = "|",
  sep = ","
)
}
\arguments{
\item{gene_cells_cluster}{The UMI cluster in each cell for a gene}

\item{filter_ratio}{The ratio of singletons to be removed.}

\item{split}{The character to seperate different exons in the isoform representation.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform representation.}

\item{splice_site_thresh}{The minimum threshold of occurance  for an splice site to be preserved}
}
\description{
correct wrong mapped and truncated reads and filter out UMI singletons.
}
\details{
correct wrong mapped and truncated reads according to other complete reads in the same UMI cluster,
and filter out UMI singletons (UMIs enriched with sequencing errors) according to UMI cluster size.
}
