% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_filter.R
\name{isoform_dis_cluster}
\alias{isoform_dis_cluster}
\title{isoform_dis_cluster}
\usage{
isoform_dis_cluster(isoforms, thresh = 20, eps = 20, split = "|", sep = ",")
}
\arguments{
\item{isoforms}{A string vector storing isoforms}

\item{thresh}{The maxmimum threshold of difference for two isoforms to be clustered}

\item{eps}{The maximum threshold of distance to be preserved}

\item{split}{The character to split each exon in the isoform representation}

\item{sep}{The character to split the start and end for each exon in the isoform representation}
}
\value{
A vector to record the cluster information of the isoforms
}
\description{
Cluster isoforms using dbscan.
}
\details{
Cluster isoforms using dbscan.
}
