% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_output.R
\name{isoformCount2ReadsForGene}
\alias{isoformCount2ReadsForGene}
\title{isoformCount2ReadsForGene}
\usage{
isoformCount2ReadsForGene(
  mat,
  genome,
  chr,
  strand,
  isoform_col = "isoform",
  count_col = "count",
  name_col = c("cell", "gene", "polyA"),
  quality = "~",
  ...
)
}
\arguments{
\item{mat}{The dataframe which records the isoform quantification, should have at least two columns, including
the isoform and the count of the isoform, there can also be other columns about the meta data, which would be saved
in the read name.}

\item{genome}{The name of the genome, will be used to retrieve the genome object from BSgenome.}

\item{chr}{The chromosome index of the sequence to be extracted.}

\item{strand}{The strand of the sequence to be extracted.}

\item{isoform_col}{The name of the column which records the isoforms}

\item{count_col}{The name of the column which records the read count}

\item{name_col}{The name of the columns which record the meta information of each read. The information from those columns would be saved in the read name.}

\item{quality}{Currently the quality for the UMI-collapsed reads are all set to the highest as '~'.}
}
\value{
A ShortReadQ object storing the fastq file.
}
\description{
Transform the isoform quantification into corresponding read sequences for a gene.
}
