% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{isomatch_penalty}
\alias{isomatch_penalty}
\title{isomatch_penalty}
\usage{
isomatch_penalty(left, right)
}
\arguments{
\item{left, right}{The end difference at the 5' and 3' end}
}
\value{
A numerical number to denote the penalty score.
}
\description{
calculate the penalty score from the difference at 5 and 3 end
}
