% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{isomatch_penalty_v}
\alias{isomatch_penalty_v}
\title{isomatch_penalty_v}
\usage{
isomatch_penalty_v(left, right)
}
\arguments{
\item{left, right}{The end difference at the 5' and 3' end}
}
\description{
The vectorized version of the isomatch_penalty function
}
