% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{isos_len}
\alias{isos_len}
\title{isos_len}
\usage{
isos_len(isos, ...)
}
\arguments{
\item{isos}{The input string of the isoform}

\item{...}{parameters for "iso_len"}
}
\value{
A numeric vector recording the length of multiple isoforms.
}
\description{
calculate the length of multiple isoform
}
