% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome.R
\name{load_genome}
\alias{load_genome}
\title{load_genome}
\usage{
load_genome(genome_name)
}
\arguments{
\item{genome_name}{The name of the genome, can be abbreviations for some
commonly used genomes like "hg38"}
}
\value{
A BSgenome object
}
\description{
Load the genome from BSgenome given the genome name
}
