% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{long2square}
\alias{long2square}
\title{long2square}
\usage{
long2square(
  long,
  row_names_from,
  col_names_from,
  values_from,
  symmetric = TRUE,
  na.fill = 0,
  nodes = NULL
)
}
\arguments{
\item{long}{The input long table}

\item{row_names_from}{The name of the column which would be transformed to each row index}

\item{col_names_from}{The name of the column which would be transformed to each col index}

\item{values_from}{The name of the column which would be set as values in the matrix}

\item{symmetric}{The flag to indicate if the transformed matrix is symmetric.}

\item{na.fill}{The value to fill the NA in the square matrix}

\item{nodes}{The pre-defined row and col index}
}
\description{
transform a long table into a square matrix
}
