% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{long2wide}
\alias{long2wide}
\title{long2wide}
\usage{
long2wide(long, row_names_from, col_names_from, values_from, symmetric = TRUE)
}
\arguments{
\item{long}{The input long table}

\item{row_names_from}{The name of the column which would be transformed to each row index}

\item{col_names_from}{The name of the column which would be transformed to each col index}

\item{values_from}{The name of the column which would be set as values in the matrix}

\item{symmetric}{The flag to indicate if the transformed matrix is symmetric.}
}
\description{
transform a long table into a wide matrix
}
