% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{louvain_iter_stack}
\alias{louvain_iter_stack}
\title{louvain_iter_stack}
\usage{
louvain_iter_stack(
  graph,
  weight = "weight",
  resolution = 1,
  alpha = 2,
  sim_thresh = 6
)
}
\arguments{
\item{graph}{an igraph object.}

\item{weight}{the name of the edge attribute to use as weight in louvain clustering.}

\item{resolution}{the resolution for the louvain clustering.}

\item{alpha}{the fold change between the number of edges and number of nodes. Lower alpha requires
highlier connected graph.}

\item{sim_thresh}{the minimum threshold for the needleman similarity for a graph to stop clustering}
}
\description{
do iterative louvain clustering for nodes in a graph
}
\details{
The function does iterative louvain clustering for an igraph object. Clusters will be
split as sub-graphs and stores in a list.
}
