% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{mid_count}
\alias{mid_count}
\title{mid_count}
\usage{
mid_count(mid, total)
}
\arguments{
\item{mid}{A string vector of middle splicing sites within a UMI cluster}
}
\value{
A dataframe, the first column is the original middle splicing sites, the second column is the chosen
representative within this UMI cluster, the third column stores the count of the original.
}
\description{
Count the frequency of middle splicing sites for each UMI cluster.
}
\details{
Count the frequency of middle splicing sites for each UMI cluster, the highest one will be chosen
as the representative and other splicing sites will be attributed to the representative.
}
