% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{mid_group}
\alias{mid_group}
\title{mid_group}
\usage{
mid_group(mid, sep = ",")
}
\arguments{
\item{mid}{A vector of strings, each string denotes the existence of the middle splicing sites}

\item{sep}{The character to seperate splicing sites in the string}
}
\value{
A dataframe with two columns, the first column denotes the middle splicing site string,
the second denotes the splicing site string that can transform to this splicing site string by
truncations.
}
\description{
group middle splicing sites which can transform to each other by truncations.
}
