% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{mid_len}
\alias{mid_len}
\title{mid_len}
\usage{
mid_len(mid, sep = ",")
}
\arguments{
\item{mid}{A vector of strings, each string denotes the existence of the middle splicing sites}

\item{sep}{The character to seperate splicing sites in the string}
}
\value{
A dataframe with two columns, the first column denotes the middle splicing sites,
the second denotes the number of sites which covered.
}
\description{
count the number of middle splicing sites covered by a read
}
