% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_extraction.R
\name{mid_polyA_filter}
\alias{mid_polyA_filter}
\title{mid_polyA_filter}
\usage{
mid_polyA_filter(endsites, genome, chr, strand, bin = 20, thresh = 0.4)
}
\arguments{
\item{endsites}{A vector of sites recording the end position for each read.}

\item{genome}{The genome annotation from BSgenome.}

\item{chr}{The chromsome of the gene to be checked.}

\item{strand}{The strand of the gene to be checked.}

\item{bin}{The window size to search for A}

\item{thresh}{The maximum threshold of the frequency of A in the search window.}
}
\value{
return bool vector to indicate if the polyA sites are real.
}
\description{
Judge if the polyA sites for a series of reads are real.
}
\details{
For the potential polyA sites in each read, this function check if the genome
sequence is enriched with A. If it is, the potential polyA site will be denied.
}
