% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{paramExtract}
\alias{paramExtract}
\title{paramExtract}
\usage{
paramExtract(func, ...)
}
\arguments{
\item{func}{functions to be match with parameters}

\item{...}{Omitted parameters}
}
\description{
Extract parameters from the ... for a function
}
\details{
Extract parameters from the ... for a function
}
