% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{paramMerge}
\alias{paramMerge}
\title{paramMerge}
\usage{
paramMerge(func, neceParam, ...)
}
\arguments{
\item{func}{functions to be match with parameters}

\item{neceParam}{A list of user input parameters}

\item{...}{Omitted parameters}
}
\description{
Merge two sets of parameters
}
\details{
Merge two sets of parameters
}
