% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_impute.R
\name{penalty2weight}
\alias{penalty2weight}
\title{penalty2weight}
\usage{
penalty2weight(x)
}
\arguments{
\item{x}{the penalty score}
}
\value{
A numerical number to denote the weights.
}
\description{
Transform the penalty score to weights for isoform alignment
}
