% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{read2bins}
\alias{read2bins}
\title{read2bins}
\usage{
read2bins(read, sep = ",", split = "|")
}
\arguments{
\item{read}{A string to represent the isoform of a read}

\item{sep}{The character to split the start and end position for each exon in the isoform.}

\item{split}{The character to split the exons in the isoform}
}
\value{
A dataframe with two columns, the first column records the start of each exon, the second records
the end position.
}
\description{
transform the read isoform string to a bin matrix
}
\details{
transform the read isoform string to a bin matrix
}
