% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_extraction.R
\name{readBam}
\alias{readBam}
\title{readBam}
\usage{
readBam(bamFile, chr, start, end, strand, map_qual = 30)
}
\arguments{
\item{bamFile}{The input bam file read by Rsamtools}

\item{chr}{The chromosome of reads located in, should be chr}

\item{start}{The start position of the reads, should be int}

\item{end}{The end position of the reads, should be int}

\item{map_qual}{The minimum mapping quality of the reads to be extracted}
}
\value{
return a list including read names, start positions, cigars and the sequences for all reads
}
\description{
extract reads in a specific region from the bam file
}
\details{
extract reads in a specific region from the bam file via Rsamtools
}
