% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{save10X}
\alias{save10X}
\title{save10X}
\usage{
save10X(long, path, i = "gene", j = "cell", value = "count")
}
\arguments{
\item{long}{The input long table}

\item{path}{The output path}

\item{i}{The name of the column which stores the features}

\item{j}{The name of the column which stores the cells}

\item{value}{The name of the column which stores the feature count}
}
\description{
save a long table into cellRanger outpur format
}
