% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{saveIsoMat}
\alias{saveIsoMat}
\title{saveIsoMat}
\usage{
saveIsoMat(
  iso,
  path,
  cell_col = "cell",
  gene_col = "gene",
  iso_col = "isoform",
  count_col = "count"
)
}
\arguments{
\item{iso}{The output from "cells_genes_isos_count", which is a dataframe with four columns}

\item{path}{The output path}

\item{cell_col}{The name of the column which stores the cells}

\item{gene_col}{The name of the column which stores the gene name}

\item{iso_col}{The name of the column which stores the isoform name}

\item{count_col}{The name of the column which stores the isoform count}
}
\description{
save the output from "cells_genes_isos_count" into 
cellRanger outpur format, for both gene quantification matrix and isoform quantification
matrix.
}
