% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{saveResult}
\alias{saveResult}
\title{saveResult}
\usage{
saveResult(data, filename)
}
\arguments{
\item{data}{A dataframe}

\item{filename}{The output filename}
}
\description{
Save the result into a file
}
\details{
Write the dataframe result in a txt table file
}
