% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{saveSparse}
\alias{saveSparse}
\title{saveSparse}
\usage{
saveSparse(data, path)
}
\arguments{
\item{data}{A matrix}

\item{path}{The output file path}
}
\description{
Save the result into a sparse matrix
}
\details{
Write the cell by gene/isoform matrix into a sparse matrix
}
