% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{site_correct}
\alias{site_correct}
\title{site_correct}
\usage{
site_correct(from, to, sep = ",")
}
\arguments{
\item{from, to}{A string vector of middle splicing sites. From is the original middle splicing
sites, to is the middle splicing sites the from mapped to within the same UMI cluster}

\item{sep}{The character to seperate splicing sites in the middle splicing site string.}
}
\value{
A datafrom recording the mapping between wrongly mapped isoform and correct isoform.
}
\description{
correct wrongly mapped splicing sites.
}
