% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_site_correct.R
\name{site_recover}
\alias{site_recover}
\title{site_recover}
\usage{
site_recover(
  start,
  end,
  mid = NULL,
  sites = NULL,
  flank = 5,
  sep = ",",
  split = "|"
)
}
\arguments{
\item{start}{A numeric value indicating start position of reads within the UMI cluster}

\item{end}{A numeric value indicating end position of reads within the UMI cluster}

\item{mid}{A string indicating patterns of middle splicing sites}

\item{sites}{A string vector indicating middle splicing sites}

\item{sep}{The character to seperate the start and end for each exon bin in the isoform representation}

\item{split}{The character to seperate each exon bin in the isoform representation}
}
\value{
A string representing isoform
}
\description{
transform the middle splicing sites pattern from boolean string into original sites
}
\details{
transform the middle splicing sites pattern from boolean string into original sites
}
