% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/reads_filter.R
\name{size_filter_error}
\alias{size_filter_error}
\title{size_filter_error}
\usage{
size_filter_error(size, ratio = 0.1)

size_filter_error(size, ratio = 0.1)
}
\arguments{
\item{size}{A vector to store the size for each cluster.}

\item{ratio}{The estimated frequency for UMIs which are enriched of sequencing errors and
couldn't be clustered correctly and become singletons}
}
\value{
A vector to indicating whether the cluster is filtered out or not

A vector to indicating whether the cluster is filtered out or not
}
\description{
Filter out singletons and small clusters which are not likely to be
correctly clustered, given the distribution of cluster size.

Filter out singletons and small clusters which are not likely to be
correctly clustered, given the distribution of cluster size.
}
