% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{SNN_graph}
\alias{SNN_graph}
\title{SNN_graph}
\usage{
SNN_graph(graph)
}
\arguments{
\item{graph}{an igraph object}
}
\description{
add neighbor sharing as an edge attribute in a graph
}
\details{
The function calculates the neighbor sharing for a graph, and add the count of shared neighbors as an edge attribute in the graph
}
