% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/splice_site_correct.R
\name{splice_site_table}
\alias{splice_site_table}
\title{splice_site_table}
\usage{
splice_site_table(isoforms, split = "|", sep = ",", splice_site_thresh = 10)

splice_site_table(isoforms, split = "|", sep = ",", splice_site_thresh = 10)
}
\arguments{
\item{split}{The character to seperate different exons in the isoform representation.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform representation.}

\item{splice_site_thresh}{The minimum number of occurance for a splice site to be preserved.}

\item{isoform}{The string which represents the isoform}
}
\value{
A dataframe, the first and the last column stores the start and end position of each read,
the middle columns store if each read has the splicing sites

A dataframe, the first and the last column stores the start and end position of each read,
the middle columns store if each read has the splicing sites
}
\description{
extract splice sites from isoform sequence.

extract splice sites from isoform sequence.
}
\details{
extract splice sites from isoform sequence and filter out artifacts.

extract splice sites from isoform sequence and filter out artifacts.
}
