% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{umi_cluster}
\alias{umi_cluster}
\title{umi_cluster}
\usage{
umi_cluster(umi, iso = NULL, thresh = NULL)
}
\arguments{
\item{umi}{a vector of UMIs.}

\item{iso}{a vector of isoform sequences.}

\item{sim_thresh}{the minimum threshold of needleman score to connect to different UMIs.}
}
\description{
apply iterative louvain clustering for UMIs.
}
\details{
apply iterative louvain clustering for UMIs based on the needleman score between each UMI. Isoform
information is optional to be used to help distinguish different UMIs.
}
