% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{UMI_consensus_out}
\alias{UMI_consensus_out}
\title{UMI_consensus_out}
\usage{
UMI_consensus_out(
  umi_count,
  dir,
  genome_path,
  genome_name,
  gene_bed,
  minimap2 = "minimap2",
  samtools = "samtools",
  minimap_bed_path = NULL,
  cores = 1,
  force_fastq_out = FALSE,
  force_map = FALSE
)
}
\arguments{
\item{umi_count}{The umi count output from umi_count_corres.}

\item{genome_path}{The filename of the reference genome}

\item{gene_bed}{The gene bed annotation}

\item{minimap2}{The path of the minimap2}

\item{samtools}{The path of the samtools}

\item{force_fastq_out}{flag to force to remap the consensused fastq}
}
\description{
Output the reads after UMI consensus into fastq and map it to the genome.
}
