% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_count_functions.R
\name{umi_count}
\alias{umi_count}
\title{umi_count}
\usage{
umi_count(
  cell_exon,
  qual,
  gene_strand,
  bar = "barcode",
  gene = "gene",
  isoform = "isoform",
  polyA = "polyA",
  sim_thresh = 7,
  split = "|",
  sep = ",",
  splice_site_thresh = 10,
  verbose = FALSE
)
}
\arguments{
\item{cell_exon}{the input dataframe which includes cell barcode, UMI sequence, isoform and polyA status.}

\item{qual}{The adapter distance from barcode matching to represent the data quality}

\item{bar}{The name of the column to represent cell barcode.}

\item{gene}{The name of column which stores the gene names}

\item{isoform}{The name of the column to represent the isoform.}

\item{polyA}{The name of the column to represent the polyA status.}

\item{sim_thresh}{the minimum threshold of needleman score to connect to different UMIs.}

\item{split}{The character to seperate different exons in the isoform representation.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform representation.}

\item{splice_site_thresh}{The minimum threshold of occurance  for an splice site to be preserved}

\item{verbose}{if print information during UMI clustering for each cell.}
}
\description{
get the umi count for each cell for each gene.
}
\details{
get the umi count for each cell for each gene, the process includes clustering
for UMI and correction for the isoform.
}
