% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{umi_count_parallel}
\alias{umi_count_parallel}
\title{umi_count_parallel}
\usage{
umi_count_parallel(
  data,
  qual,
  dir,
  gene_bed,
  bar = "barcode",
  gene = "gene",
  isoform = "isoform",
  polyA = "polyA",
  sim_thresh = NULL,
  split = "|",
  sep = ",",
  splice_site_thresh = 3,
  verbose = FALSE,
  bed_gene_col = "gene",
  bed_strand_col = "strand",
  cores = 1,
  force_UMI_dedup = FALSE
)
}
\arguments{
\item{data}{The first output dataframe from reads_extract_bc.}

\item{qual}{The adapter distance from barcode matching to represent the data quality}

\item{bar}{The name of the column to represent cell barcode.}

\item{gene}{The name of column which stores the gene names}

\item{isoform}{The name of the column to represent the isoform.}

\item{polyA}{The name of the column to represent the polyA status.}

\item{sim_thresh}{the minimum threshold of needleman score to connect to different UMIs.}

\item{split}{The character to seperate different exons in the isoform representation.}

\item{sep}{The character to seperate the start and end position of an exon in the isoform representation.}

\item{splice_site_thresh}{The minimum threshold of occurance  for an splice site to be preserved}

\item{verbose}{if print information during UMI clustering for each cell.}

\item{cores}{The number of cores required for parallel}

\item{force_UMI_dedup}{flag to force to redo the UMI deduplication step}

\item{force_fastq_out}{flag to force to redo the fastq output step}
}
\description{
A wrapper function to include all steps in the UMI deduplication.
}
\details{
A wrapper function to include all steps in the UMI deduplication, including UMI cluster,
isoform correction, isoform imputation and quantification.
}
