% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi_cluster.R
\name{umi_sim_graph}
\alias{umi_sim_graph}
\title{umi_sim_graph}
\usage{
umi_sim_graph(
  umi,
  iso = NULL,
  sim_thresh = 7,
  iso_thresh = 200,
  split = "|",
  sep = ","
)
}
\arguments{
\item{umi}{The umi sequence for each read}

\item{iso}{The isoform sequence for each read}

\item{sim_thresh}{The minimum threshold of the similarity. Only UMIs with higher similarity will be connected with an edge.}

\item{iso_thresh}{The maximum threshold that two isoforms can be different in overlapping bases.}

\item{split}{The character to split each exon in the isoform representation}

\item{sep}{The character to seperate the exon start and end in the isoform representation}
}
\description{
build UMI graph based on their sequence similarity
}
\details{
The function calculates the similarity of two UMIs by needleman score. Their isoform similarity will also be considered if given.
Based on the similarity a sNN graph will be built.
}
