% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_extraction.R
\name{extractTagBc}
\alias{extractTagBc}
\title{extractTagBc}
\usage{
extractTagBc(
  fastq_path,
  barcode_path,
  out_name,
  toolkit,
  protocol = "10X",
  adapter = NULL,
  window = 10,
  step = 2,
  left_flank = 0,
  right_flank = 0,
  drop_adapter = FALSE,
  polyA_bin = 20,
  polyA_base_count = 15,
  polyA_len = 10,
  barcode_len = 16,
  mu = 20,
  sigma = 10,
  k = 6,
  batch = 100,
  top = 5,
  cos_thresh = 0.25,
  alpha = 0.05,
  edit_thresh = 3,
  mean_edit_thresh = 1.5,
  UMI_len = 10,
  UMI_flank = 1,
  cores = 1
)
}
\arguments{
\item{fastq_path}{The path of the input fastq}

\item{barcode_path}{The path for the barcode whitelist}

\item{out_name}{The path for the polished fastq}

\item{toolkit}{The kit to build the library, should be 5 or 3}

\item{adapter}{The sequence of the adapter which is aside the cell barcode}

\item{window}{The window size to search the substring of adapter in the read}

\item{step}{The step size to search the substring of adapter in the read}

\item{left_flank, right_flank}{The length of the left/right part aside the adapter to be extracted}

\item{drop_adapter}{A flag to indicate if the adapter region should be removed from the tag region.}

\item{polyA_bin}{The window size to search for polyA.}

\item{polyA_base_count}{The minimum threshold of the number of A within the polyA search window}

\item{polyA_len}{The maximum length of polyA to be preserved in the read}

\item{mu}{The initial expectation of the start positions for cell barcodes in the tag region}

\item{sigma}{The initial variance of the start positions for cell barcodes in the tag region}

\item{k}{The length of barcode substring search in the tag region}

\item{batch}{The number of reads to be processed to update the distribution of
start position of cell barcodes}

\item{top}{The number of candidate barcodes which are further compared with edit distance}

\item{cos_thresh}{The minimum threshold for the cosine similarity between the barcode and
the tag region, only barcode with cosine similarity larger than the thresh can be preserved as
candidate barcodes.}

\item{alpha}{The size of the confidence interval for the start position of cell barcode alignment}

\item{edit_thresh}{The maximum threshold for the edit distance of the barcode alignment, alighment with
edit distance over the threshold would be filtered out.}

\item{UMI_len}{The length of the UMI}

\item{cores}{The number of cores to use for parallization}

\item{flank}{The length of flank when extract UMI, which is used to be tolerant of indels and dels.}
}
\description{
Extract the tag region and identify cell barcode in the ta region
}
\details{
Extract the tag region for each reads and trim the tag region out to generate a
polished read. Then the tag region is used to identify cell barcode an UMI.
}
