% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{reads_extract_bc}
\alias{reads_extract_bc}
\title{reads_extract_bc}
\usage{
reads_extract_bc(
  fastq_path,
  barcode_path,
  gene_bed,
  adapter,
  genome_path,
  genome_name,
  toolkit,
  protocol = "10X",
  minimap_bed_path = NULL,
  work_dir = "./",
  window = 10,
  step = 2,
  left_flank = 0,
  right_flank = 0,
  drop_adapter = FALSE,
  polyA_bin = 20,
  polyA_base_count = 15,
  polyA_len = 10,
  barcode_len = 16,
  mu = 15,
  sigma = 10,
  k = 6,
  batch = 100,
  top = 5,
  cos_thresh = 0.25,
  alpha = 0.05,
  edit_thresh = 3,
  UMI_len = 10,
  UMI_flank = 1,
  minimap2 = "minimap2",
  samtools = "samtools",
  bedtools = "bedtools",
  map_qual = 30,
  end_flank = 200,
  splice_site_bin = 2,
  mean_edit_thresh = 1.5,
  force_barcode_match = FALSE,
  force_map = FALSE,
  force_isoform_extract = FALSE,
  force_rerun = FALSE,
  cores = 1
)
}
\arguments{
\item{fastq_path}{The path of the input fastq}

\item{barcode_path}{The path for the barcode whitelist}

\item{gene_bed}{The gene bed annotation, should be a dataframe}

\item{adapter}{The sequence of the adapter which is aside the cell barcode}

\item{genome_path}{The filename of the reference genome}

\item{toolkit}{The kit to build the library, should be 5 or 3}

\item{work_dir}{The folder to store intermediate files and output}

\item{window}{The window size to search the substring of adapter in the read}

\item{step}{The step size to search the substring of adapter in the read}

\item{left_flank, right_flank}{The length of the left/right part aside the adapter to be extracted}

\item{drop_adapter}{A flag to indicate if the adapter region should be removed from the tag region.}

\item{polyA_bin}{The window size to search for polyA.}

\item{polyA_base_count}{The minimum threshold of the number of A within the polyA search window}

\item{polyA_len}{The maximum length of polyA to be preserved in the read}

\item{barcode_len, UMI_len}{The length of the cell barcode/UMI}

\item{mu}{The initial expectation of the start positions for cell barcodes in the tag region}

\item{sigma}{The initial variance of the start positions for cell barcodes in the tag region}

\item{k}{The length of barcode substring search in the tag region}

\item{batch}{The number of reads to be processed to update the distribution of
start position of cell barcodes}

\item{top}{The number of candidate barcodes which are further compared with edit distance}

\item{cos_thresh}{The minimum threshold for the cosine similarity between the barcode and
the tag region, only barcode with cosine similarity larger than the thresh can be preserved as
candidate barcodes.}

\item{alpha}{The size of the confidence interval for the start position of cell barcode alignment}

\item{edit_thresh}{The maximum threshold for the edit distance of the barcode alignment, alighment with
edit distance over the threshold would be filtered out.}

\item{minimap2}{The path of the minimap2}

\item{samtools}{The path of the samtools}

\item{map_qual}{The minimum mapping quality of the reads to be extracted}

\item{end_flank}{The maximum threshold for a read to exceed the end of the gene region annotation}

\item{splice_site_bin}{The bin size to correct slice sites given the splice sites annotation from the gene bed}

\item{mean_edit_thresh}{The maximum tolerance of mean edit distance for a cell barcode to be preserved}

\item{force_barcode_match}{flag to force to redo the barcode match step}

\item{force_map}{flag to force to redo the mapping step}

\item{force_isoform_extract}{flag to force to redo the isoform extraction step}

\item{cores}{The number of cores to use for parallization}
}
\value{
A list with two dataframes: the first one records the barcode, UMI and isoform information
for each read, the second one stores the needleman score distirbution for adapters as evaluation of
data quality.
}
\description{
A wrapper function to include extractTagBc, fastqMap and reads_extraction
}
\details{
A wrapper function to include extractTagBc, fastqMap and reads_extraction. And this function
would integrate the output from above functions by read name.
}
