% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_extraction.R
\name{reads_extraction}
\alias{reads_extraction}
\title{reads_extraction}
\usage{
reads_extraction(
  bam_path,
  gene_bed,
  genome,
  toolkit = 5,
  map_qual = 30,
  end_flank = 200,
  splice_site_bin = 2,
  mid_polyA_bin = 20,
  mid_polyA_thresh = 0.4
)
}
\arguments{
\item{bam_path}{The path of the input bam file}

\item{gene_bed}{The gene bed annotation, should be a dataframe}

\item{genome}{The genome annotation from BSgenome.}

\item{toolkit}{The position of cell barcode and UMI, should only be 5 or 3 end for a library}

\item{map_qual}{The minimum mapping quality of the reads to be extracted}

\item{end_flank}{The maximum threshold for a read to exceed the end of the gene region annotation}

\item{splice_site_bin}{The bin size to correct slice sites given the splice sites annotation from the gene bed}

\item{mid_polyA_bin}{The window size to search for A}

\item{mid_polyA_thresh}{The maximum threshold of the frequency of A in the search window.}

\item{cores}{The number of cores for parallization}
}
\value{
A dataframe including the exons, and polyA existence, each row is a read.
}
\description{
extract reads for multiple genes from the bam given the gene bed annotation
}
\details{
extract reads information for a gene from the bam given the gene bed annotation, including
exon bins and polyA existence
}
