% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{RunLongcellPre}
\alias{RunLongcellPre}
\title{RunLongcellPre}
\usage{
RunLongcellPre(
  fastq_path,
  barcode_path,
  toolkit,
  genome_path,
  genome_name,
  protocol = "10X",
  adapter = NULL,
  gtf_path = NULL,
  gene_bed_path = NULL,
  minimap_bed_path = NULL,
  to_isoform = TRUE,
  work_dir = "./",
  cores = 1,
  mode = c("sequential", "multisession", "multicore", "cluster"),
  minimap2 = "minimap2",
  samtools = "samtools",
  bedtools = "bedtools",
  ...
)
}
\arguments{
\item{fastq_path}{The path of the input fastq}

\item{barcode_path}{The path for the barcode whitelist}

\item{toolkit}{The kit to build the library, should be 5 or 3}

\item{genome_path}{The filename of the reference genome}

\item{adapter}{The sequence of the adapter which is aside the cell barcode}

\item{work_dir}{The folder to store intermediate files and output}

\item{cores}{The number of cores to use for parallization}

\item{minimap2}{The path of the minimap2}

\item{samtools}{The path of the samtools}

\item{...}{All other tuning parameters}

\item{force_barcode_match}{decide whether to redo the barcode match if there already exist the output}
}
\description{
A wrapper function to include the reads_extract_bc and umi_count_parallel.
}
\details{
A wrapper function to include the reads_extract_bc and umi_count_parallel.
}
