% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_extraction.R
\name{seq_config}
\alias{seq_config}
\title{seq_config}
\usage{
seq_config(protocol, toolkit)
}
\arguments{
\item{protocol}{The sequence protocol}

\item{toolkit}{The kit to build the library, should be 5 or 3}
}
\description{
Preset parameters for different sequencing protocol
}
